backend/requirements.txt
pandas==2.0.1
boto3==1.28.34
selenium==4.15.0
docker==7.0.0
numpy==1.26.3
python-dotenv==1.0.0
openpyxl==3.1.2
ruff==0.9.6
